feat: clean up structure of component and remove uneeded logic
