--- packages/date-picker/src/Calendar.scss
@@ -1,107 +1,0 @@
-@import "~@kaizen/design-tokens/sass/typography";
-@import "~@kaizen/design-tokens/sass/color";
-@import "~@kaizen/design-tokens/sass/spacing";
-@import "~@kaizen/design-tokens/sass/border";
-@import "~@kaizen/draft-button/KaizenDraft/Button/styles";
-
-// Vars
-$border-radius: 3px;
-
-.calendar {
-  font-family: $typography-heading-4-font-family;
-
-  .wrapper {
-    padding-bottom: 0;
-  }
-
-  .navbar {
-    position: absolute;
-    display: flex;
-    justify-content: space-between;
-    width: calc(100% - 3rem);
-    margin: 0 $spacing-md;
-    top: 1.25rem;
-    color: $color-purple-800;
-
-    .arrows {
-      @include button-reset;
-      display: inherit;
-      padding: 0.25rem $spacing-xs;
-      border: $border-solid-border-width $border-solid-border-style transparent;
-      border-radius: $border-radius;
-
-      &:hover {
-        cursor: pointer;
-        background-color: $color-gray-100;
-      }
-
-      &:focus {
-        border: $color-blue-500 $border-focus-ring-border-style
-          $border-focus-ring-border-width;
-      }
-    }
-  }
-
-  .month {
-    margin: $spacing-md;
-    border-collapse: separate;
-    border-spacing: 1px;
-  }
-
-  .months {
-    box-shadow: 0px 0px 7px rgb(0 0 0 / 9%);
-    border-radius: $border-solid-border-radius;
-  }
-
-  .caption {
-    text-align: center;
-    font-size: $typography-heading-4-font-size;
-    font-weight: $typography-heading-4-font-weight;
-    line-height: $typography-heading-4-line-height;
-    letter-spacing: $typography-heading-4-letter-spacing;
-    color: $color-purple-800;
-    margin: 0 0 $spacing-xs;
-  }
-
-  .weekday {
-    font-size: $typography-paragraph-body-font-size;
-    font-weight: $typography-paragraph-body-font-weight;
-    line-height: $typography-paragraph-body-line-height;
-    color: $color-purple-800;
-  }
-
-  .body {
-    font-size: $typography-paragraph-body-font-size;
-    font-weight: $typography-paragraph-body-font-weight;
-    line-height: $typography-paragraph-body-line-height;
-  }
-
-  .today {
-    color: $color-blue-500;
-  }
-
-  .day {
-    padding: calc(#{$spacing-md} / 3) 0.625rem;
-    border-radius: $border-radius;
-
-    &:focus-visible {
-      outline-offset: -1px;
-    }
-  }
-
-  .disabled {
-    color: rgba($color-purple-800-rgb, 0.3);
-  }
-
-  .day:not(.interactionDisabled):not(.disabled):not(.selected):not(.outside):hover {
-    border-radius: $border-radius;
-    background-color: $color-blue-100;
-    color: $color-blue-500;
-  }
-
-  .selected:not(.disabled):not(.outside) {
-    border-radius: $border-radius;
-    background-color: $color-blue-500;
-    color: $color-white;
-  }
-}
