+++ packages/date-picker/src/DatePicker.tsx
@@ -1,14 +1,11 @@
 import { TextField } from "@kaizen/draft-form"
-import { Icon } from "@kaizen/component-library"
-import React, { useEffect, useRef, useState } from "react"
+import React, { useState } from "react"
 import dateStart from "@kaizen/component-library/icons/date-start.icon.svg"
-import arrowRight from "@kaizen/component-library/icons/arrow-right.icon.svg"
-import arrowLeft from "@kaizen/component-library/icons/arrow-left.icon.svg"
-import classnames from "classnames"
 import "react-day-picker/lib/style.css"
 import DayPicker from "react-day-picker/DayPicker"
 import styles from "./DatePicker.scss"
-import calendarStyles from "./Calendar.scss"
+import { CalendarNav, CalendarNavProps } from "./CalendarNav"
+import { defaultDatePickerClasses } from "./DatePickerClasses"
 
 type DatePickerProps = {
   selectedDate?: Date
@@ -23,14 +20,9 @@ export const DatePickerWrapper: React.FunctionComponent<DatePickerProps> = ({
   labelText,
   isDisabled = false,
 }) => {
-  const getNavbar = ({ ...navbarProps }: NavbarProps) => (
-    <Navbar {...navbarProps} />
+  const getNavbar = ({ ...navbarProps }: CalendarNavProps) => (
+    <CalendarNav {...navbarProps} />
   )
-
-  const [showPicker, setShowPicker] = useState(false)
-  const handleFocus = () => {
-    setShowPicker(true)
-  }
 
   return (
     <>
@@ -38,74 +30,23 @@ export const DatePickerWrapper: React.FunctionComponent<DatePickerProps> = ({
         id="date-picker"
         labelText={labelText}
         icon={dateStart}
-        onFocus={handleFocus}
-        inputValue={selectedDate}
+        inputValue={
+          selectedDate ? selectedDate.toLocaleDateString() : undefined
+        }
         disabled={isDisabled}
       />
 
       <DayPicker
         selectedDays={selectedDate}
+        // TODO create disabledDays prop, this is to display disabled styling
         disabledDays={{
           daysOfWeek: [6],
         }}
         onDayClick={onDayChange}
         navbarElement={getNavbar}
         classNames={defaultDatePickerClasses}
-        className={calendarStyles.calendar}
+        className={styles.calendar}
       />
     </>
   )
 }
-
-export const defaultDatePickerClasses = {
-  container: "DayPicker",
-  wrapper: `${calendarStyles.wrapper} DayPicker-wrapper`,
-  interactionDisabled: `${calendarStyles.interactionDisabled} DayPicker--interactionDisabled`,
-  months: `${calendarStyles.months} DayPicker-Months`,
-  month: `${calendarStyles.month} DayPicker-Month`,
-  navBar: "DayPicker-NavBar",
-  navButtonPrev: "DayPicker-NavButton DayPicker-NavButton--prev",
-  navButtonNext: "DayPicker-NavButton DayPicker-NavButton--next",
-  navButtonInteractionDisabled: "DayPicker-NavButton--interactionDisabled",
-  caption: `${calendarStyles.caption} DayPicker-Caption`,
-  weekdays: "DayPicker-Weekdays",
-  weekdaysRow: "DayPicker-WeekdaysRow",
-  weekday: `${calendarStyles.weekday} DayPicker-Weekday`,
-  body: `${calendarStyles.body} DayPicker-Body`,
-  week: "DayPicker-Week",
-  weekNumber: "DayPicker-WeekNumber",
-  day: `${calendarStyles.day} DayPicker-Day`,
-  footer: "DayPicker-Footer",
-  todayButton: "DayPicker-TodayButton",
-  today: `${calendarStyles.today} DayPicker-Day--today`,
-  selected: `${calendarStyles.selected} DayPicker-Day--selected`,
-  disabled: `${calendarStyles.disabled} DayPicker-Day--disabled`,
-  outside: `${calendarStyles.outside} DayPicker-Day--outside`,
-}
-
-type NavbarProps = {
-  onPreviousClick: () => void
-  onNextClick: () => void
-}
-
-const Navbar: React.FunctionComponent<NavbarProps> = ({
-  onPreviousClick,
-  onNextClick,
-}) => (
-  <div className={calendarStyles.navbar}>
-    <button
-      className={calendarStyles.arrows}
-      onClick={() => onPreviousClick()}
-      type="button"
-    >
-      <Icon icon={arrowLeft} role="presentation" />
-    </button>
-    <button
-      className={calendarStyles.arrows}
-      onClick={() => onNextClick()}
-      type="button"
-    >
-      <Icon icon={arrowRight} role="presentation" />
-    </button>
-  </div>
-)
