+++ packages/date-picker/src/DatePickerClasses.ts
@@ -1,0 +1,27 @@
+import styles from "./DatePicker.scss"
+
+export const defaultDatePickerClasses = {
+  container: "DayPicker",
+  wrapper: `${styles.wrapper} DayPicker-wrapper`,
+  interactionDisabled: `${styles.interactionDisabled} DayPicker--interactionDisabled`,
+  months: `${styles.months} DayPicker-Months`,
+  month: `${styles.month} DayPicker-Month`,
+  navBar: "DayPicker-NavBar",
+  navButtonPrev: "DayPicker-NavButton DayPicker-NavButton--prev",
+  navButtonNext: "DayPicker-NavButton DayPicker-NavButton--next",
+  navButtonInteractionDisabled: "DayPicker-NavButton--interactionDisabled",
+  caption: `${styles.caption} DayPicker-Caption`,
+  weekdays: "DayPicker-Weekdays",
+  weekdaysRow: "DayPicker-WeekdaysRow",
+  weekday: `${styles.weekday} DayPicker-Weekday`,
+  body: `${styles.body} DayPicker-Body`,
+  week: "DayPicker-Week",
+  weekNumber: "DayPicker-WeekNumber",
+  day: `${styles.day} DayPicker-Day`,
+  footer: "DayPicker-Footer",
+  todayButton: "DayPicker-TodayButton",
+  today: `${styles.today} DayPicker-Day--today`,
+  selected: `${styles.selected} DayPicker-Day--selected`,
+  disabled: `${styles.disabled} DayPicker-Day--disabled`,
+  outside: `${styles.outside} DayPicker-Day--outside`,
+}
